data_generator/requirements.txt
paho-mqtt
json-logging